project_doc_analyzer/requirements.txt
streamlit
langchain
openai
faiss-cpu
python-docx
pypdf
unstructured
tiktoken
python-dotenv
